Remove send message from user
