+++ apps/io-func-sign-issuer/src/app/use-cases/send-notification.ts
@@ -68,7 +68,7 @@ export const makeSendNotification =
       getDossier,
       requestToSignMessage
     );
-    pipe(
+    return pipe(
       sequenceS(TE.ApplySeq)({
         signatureRequest: pipe(
           signatureRequest,
